+++ DualDistrictStorage/DualDistrictStorageSpec.cs
@@ -1,0 +1,5 @@
+using Timberborn.BlueprintSystem;
+
+namespace IgorZ.DualDistrictStorage;
+
+sealed record DualDistrictStorageSpec : ComponentSpec;