+++ DualDistrictStorage/ResourceCountingServicePatch.cs
@@ -1,14 +1,14 @@
 using HarmonyLib;
 using Timberborn.ResourceCountingSystem;
 
-namespace IgorZ.DualDistrictWarehouse;
+namespace IgorZ.DualDistrictStorage;
 
 [HarmonyPatch(typeof(ResourceCountingService))]
 static class ResourceCountingServicePatch {
   [HarmonyPostfix]
   [HarmonyPatch(nameof(ResourceCountingService.GetGlobalResourceCount))]
   static void GetGlobalResourceCount(string goodId, ref ResourceCount __result) {
-    var registry = DualDistrictWarehouseRegistry.Instance;
+    var registry = DualDistrictStorageRegistry.Instance;
     if (registry != null) {
       __result = registry.Deduplicate(goodId, __result);
     }