+++ DualDistrictWarehouse/DualDistrictWarehouseSpec.cs
@@ -1,5 +1,0 @@
-using Timberborn.BlueprintSystem;
-
-namespace IgorZ.DualDistrictWarehouse;
-
-sealed record DualDistrictWarehouseSpec : ComponentSpec;